nevermind. auth routes will be in index, was just missing a /
